# src/docs/swagger.yaml
definitions:
  models.Response:
    properties:
      code:
        description: 业务状态码
        example: 0
        type: integer
      data:
        description: 数据体
        type: object
      message:
        description: 异常信息
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: 任务管理系统
  license: {}
  title: 任务管理 API
  version: "1.0"
paths:
  /v1/job:
    get:
      description: 获取所有任务
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取所有任务
      tags:
      - v1/job
    post:
      consumes:
      - multipart/form-data
      description: 新建任务
      parameters:
      - description: 任务名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Response'
      summary: 新建任务
      tags:
      - v1/job
  /v1/job/{id}:
    delete:
      description: 删除任务
      parameters:
      - description: 任务id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Response'
      summary: 删除任务
      tags:
      - v1/job
    get:
      description: 获取单个任务
      parameters:
      - description: 任务id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取单个任务
      tags:
      - v1/job
    put:
      consumes:
      - multipart/form-data
      description: 更新任务
      parameters:
      - description: 任务id
        in: path
        name: id
        required: true
        type: integer
      - description: 任务结果
        in: formData
        name: result
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Response'
      summary: 更新任务
      tags:
      - v1/job
swagger: "2.0"
tags:
- description: 任务管理
  name: v1/job